requirements.txt
joblib==0.16.0
librosa==0.8.0
scipy==1.5.2
PyYAML==5.1.2
tqdm==4.43.0
numpy==1.16.4
pandas==0.23.4
tensorboardX==1.9
torch==1.6.0
matplotlib==3.3.0
Pillow>=6.2.2
numba==0.48